refactor: move config-related utilities to config directory
