+++ packages/broker/src/config/ConfigWizard.ts
@@ -6,9 +6,9 @@ import chalk from 'chalk'
 import { v4 as uuid } from 'uuid'
 import * as Protocol from 'streamr-client-protocol'
 
-import * as WebsocketConfigSchema from './plugins/websocket/config.schema.json'
-import * as MqttConfigSchema from './plugins/mqtt/config.schema.json'
-import * as BrokerConfigSchema from './helpers/config.schema.json'
+import * as WebsocketConfigSchema from '../plugins/websocket/config.schema.json'
+import * as MqttConfigSchema from '../plugins/mqtt/config.schema.json'
+import * as BrokerConfigSchema from '../helpers/config.schema.json'
 import { getDefaultFile } from './config'
 
 const createLogger = () => {
