+++ packages/broker/src/config/validateConfig.ts
